+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Documentation updates.
 
+- Update to v2.5.0 of the Google Beta Terraform Provider.
+
 ## 0.18.3 (Released April 5, 2019)
 
 - Fixed issue where picking certain names for `cloudfunction.Function`s or `storage.Bucket`s would